+++ eras/mary/impl/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-mary`
 
-## 1.5.0.1
+## 1.5.1.0
 
-*
+* Add implementation for `getMinFeeTxUtxo`
 
 ## 1.5.0.0